fixes #8: enhances the "searches without results" report
enhances the "searches without results" report to give a better idea of how often that search string is searched.  instead of a log of searches, which can be seen on the "search log" report, it shows all search strings that have returned no results, and the number of times that search string was searched.  this way you can quickly tell if 100 different searches for "Widget" all returned no results.  if only one search was performed, its not a big deal, but if lots of searches for that string were performed then you will want to address your search indexing to help users find the content they are looking for.