fixes #36: tagStripper() does not remove HTML comments
since we are using tagStripper() to remove HTML from text sent to Solr for the highlighting field, we need to make sure there are no HTML comments in the text sent as well.  tagStripper() did not support removing HTML comments, so it has been modified to do so.  this should prevent a highlighting section from container a start html comment but no end which cause the display to break.